requirements.txt
# Tinker Framework
tinker>=0.1.0
tinker-cookbook

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
datasets>=2.14.0
huggingface-hub>=0.16.0

# NLP and Tokenization
transformers>=4.30.0
tokenizers>=0.13.0
sentencepiece>=0.1.99

# Romanian Language Processing
nltk>=3.8.0
langdetect>=1.0.9

# Data Quality and Validation
jsonschema>=4.17.0
tqdm>=4.65.0

# Translation (for data augmentation if needed)
deep-translator>=1.11.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0

# Jupyter for exploration
jupyter>=1.0.0
ipywidgets>=8.0.0

# Evaluation and Metrics
scikit-learn>=1.3.0
rouge-score>=0.1.2

# Logging and Monitoring
wandb>=0.15.0  # Optional: for experiment tracking
